# Conflicts:
#	src/ru/koylubaevaem/task2git/SalesManager.java

+++ src/ru/koylubaevaem/task2git/SalesManager.java
@@ -2,15 +2,15 @@ package ru.koylubaevaem.task2git;
 
 public class SalesManager {
 
-    protected int[] sales;
+    protected long[] sales;
 
-    public SalesManager(int[] sales) {
+    public SalesManager(long[] sales) {
         this.sales = sales;
     }
 
-    public int max() {
-        int max = sales[0];
-        for (int sale : sales) {
+    public long max() {
+        long max = sales[0];
+        for (long sale : sales) {
             if (sale > max) {
                 max = sale;
             }
@@ -19,10 +19,10 @@ public class SalesManager {
     }
 
     public double average() {
-        int min = sales[0];
-        int max = sales[0];
+        long min = sales[0];
+        long max = sales[0];
         for (int i = 1; i < sales.length; i++) {
-            int sale = sales[i];
+            long sale = sales[i];
             if (sale < min) {
                 min = sale;
             }
@@ -32,7 +32,7 @@ public class SalesManager {
         }
         int sum = 0;
         int count = 0;
-        for (int sale : sales) {
+        for (long sale : sales) {
             if (sale != min && sale != max) {
                 sum += sale;
                 count++;
